--- app/(routes)/layout.tsx
@@ -1,7 +1,0 @@
-export default function RootLayout({
-	children,
-}: Readonly<{
-	children: React.ReactNode;
-}>) {
-	return <>{children}</>;
-}
