+++ app/layout.tsx
@@ -16,8 +16,10 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
 	children,
+	modals,
 }: Readonly<{
 	children: React.ReactNode;
+	modals: React.ReactNode;
 }>) {
 	return (
 		<html lang="en">
@@ -29,6 +31,7 @@ export default function RootLayout({
 			>
 				<Navbar />
 				<main className="overflow-hidden relative">
+					{modals}
 					{children}
 				</main>
 				<Footer />
